+++ ui/src/pages/Colors/ColorPicker/constant.ts
@@ -1,0 +1,2 @@
+const formatLabels = ["HEX", "RGBA", "RGB", "HSL", "HSLA"] as const;
+export { formatLabels };